% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripser-dist.R
\name{ripser_dist}
\alias{ripser_dist}
\title{Calculate Persistent Homology via a Vietoris-Rips Filtration}
\usage{
ripser_dist(dataset, max_dim = 1L, threshold = Inf)
}
\arguments{
\item{dataset}{"dist" object on which to compute persistent homology}

\item{max_dim}{maximum dimension}

\item{threshold}{maximum diameter}
}
\value{
an empty list
}
\description{
This "externally exported" R function gives the user access to
  the "internally exported" C++ function `ripser_cpp_dist()`.
}
\examples{

# validate computation on toy data set
dist_vec <- c(4, 3, 5, 5, 3, 4)
result <- ripserq:::ripser_cpp_dist(
  dist_vec,
  dim = 1,
  thresh = 6.0,
  ratio = 1.0,
  p = 2
)
result

# validate clustering (only 0-degree homology)
ripserq:::ripser_cpp_dist(
  dist_vec,
  dim = 0,
  thresh = Inf,
  ratio = 1.0,
  p = 2
)

# validate use of default threshold
ripserq:::ripser_cpp_dist(
  dist_vec,
  dim = 1,
  thresh = Inf,
  ratio = 1.0,
  p = 2
)

# validate compatibility with 'dist' class
ripserq:::ripser_cpp_dist(
  eurodist,
  dim = 1, thresh = 5000, ratio = 1.0, p = 2
)

# exposed R function with no explicit parameter settings
ripser_dist(dist_vec)

}
